backend/requirements.txt
fastapi
uvicorn
praw
python-jose[cryptography]
passlib[bcrypt]
python-multipart
motor
pydantic[email]
cryptography